requirements.txt
python-telegram-bot[job-queue]==21.0.1
python-dotenv==1.0.1
requests==2.31.0
supabase==2.3.6
httpx==0.27.0
